minor cleanup of collective sources
	modified:   .github/workflows/ubuntu-clang.yml
	modified:   .github/workflows/ubuntu-gcc.yml
	modified:   CMakeLists.txt
	modified:   src/synergia/collective/CMakeLists.txt
	modified:   src/synergia/collective/deposit.cc
	modified:   src/synergia/collective/impedance.cc
	modified:   src/synergia/collective/space_charge_2d_open_hockney.cc
	modified:   src/synergia/collective/space_charge_3d_fd.cc
	modified:   src/synergia/collective/space_charge_3d_fd_impl.h
	deleted:    src/synergia/collective/space_charge_3d_fd_kernels.h
	modified:   src/synergia/collective/space_charge_3d_fd_utils.cc
	modified:   src/synergia/collective/space_charge_3d_fd_utils.h
	new file:   src/synergia/collective/space_charge_3d_kernels.h
	modified:   src/synergia/collective/space_charge_3d_open_hockney.cc
	modified:   src/synergia/collective/space_charge_rectangular.cc
	modified:   src/synergia/collective/utils.h
	modified:   src/synergia/utils/CMakeLists.txt
	renamed:    src/synergia/utils/kokkos_tools.h -> src/synergia/utils/kokkos_utils.h
