+++ .github/workflows/ubuntu-clang.yml
@@ -100,4 +100,4 @@ jobs:
         export LD_LIBRARY_PATH=${SYNINSTALL}/lib:${SYNINSTALL}/lib64:\${LD_LIBRARY_PATH}
         export PYTHONPATH=${SYNINSTALL}/lib:${SYNINSTALL}/lib/python3.9/site-packages:\${PYTHONPATH}
         cd ${{github.workspace}}/build/examples/fodo
-        mpirun -np 1 python3.9 fodo_readme.py
+        mpirun -np 1 python fodo_readme.py
