+++ CMakeLists.txt
@@ -6,10 +6,6 @@ endif()
 
 project(SYNERGIA2 LANGUAGES CXX C)
 
-if(${CMAKE_VERSION} VERSION_GREATER "3.22.0")
-  cmake_policy(SET CMP0127 OLD) # remove this when we move to CMake 3.22
-endif()
-
 # generate compile_commands.json
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 # and place it in the source directory https://stackoverflow.com/a/60910583
